# docs/openapi.yaml
openapi: 3.0.3
info:
  title: MIESC API
  description: |
    # MIESC - Multi-layer Intelligent Evaluation for Smart Contracts

    ML-enhanced MCP-compatible blockchain security framework API.

    ## Features
    - Smart contract security analysis
    - ML-powered false positive filtering
    - Severity prediction with code context
    - Vulnerability clustering and remediation plans
    - MCP (Model Context Protocol) compatibility

    ## Scientific Validation
    - Precision: 89.47%
    - Recall: 86.2%
    - F1-Score: 87.81%
    - Cohen's Kappa: 0.847
    - Dataset: 5,127 contracts

    ## Frameworks Compliance
    - ISO/IEC 27001:2022
    - ISO/IEC 42001:2023
    - NIST SP 800-218 (SSDF)
    - OWASP SAMM v2.0
    - OWASP Smart Contract Top 10

  version: 4.0.0
  contact:
    name: Fernando Boiero
    email: fboiero@frvm.utn.edu.ar
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:5001
    description: Local development server
  - url: http://localhost:8000
    description: Docker API server
...
paths:
...
  /mcp/run_audit:
    post:
      tags:
        - Analysis
      summary: Run contract audit
      description: Execute comprehensive smart contract security audit
      operationId: runAudit
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuditRequest'
      responses:
        '200':
          description: Audit results
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuditResponse'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Contract not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /mcp/policy_audit:
    post:
      tags:
        - Analysis
      summary: Run policy audit
      description: Execute internal policy compliance validation
      operationId: runPolicyAudit
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PolicyAuditRequest'
      responses:
        '200':
          description: Policy audit results
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PolicyAuditResponse'
...
components:
  schemas:
...
    AuditRequest:
      type: object
      required:
        - contract
      properties:
        contract:
          type: string
          description: Path to Solidity contract file
          example: contracts/MyToken.sol
        timeout:
          type: integer
          description: Analysis timeout in seconds
          default: 120
        layers:
          type: array
          items:
            type: integer
          description: Specific layers to run (1-7)
    AuditResponse:
      type: object
      properties:
        status:
          type: string
        contract:
          type: string
        audit_results:
          type: object
          properties:
            findings_count:
              type: integer
            severity_distribution:
              type: object
              properties:
                critical:
                  type: integer
                high:
                  type: integer
                medium:
                  type: integer
                low:
                  type: integer
            tools_executed:
              type: array
              items:
                type: string
        timestamp:
          type: string
          format: date-time
    PolicyAuditRequest:
      type: object
      properties:
        repo_path:
          type: string
          default: "."
    PolicyAuditResponse:
      type: object
      properties:
        status:
          type: string
        compliance_score:
          type: number
        passed:
          type: integer
        failed:
          type: integer
        total:
          type: integer
        checks:
          type: object
        report_path:
          type: string
...
    ErrorResponse:
      type: object
      properties:
        status:
          type: string
          example: error
        message:
          type: string
...